refactor: fanout msg の実行層を internal/app/peermsg へ分解 (layered 7/10)
cmd/fanout/msg.go(1,054 行)を CLI 境界(parse/usage、379 行残置)と
実行層(identity 解決・DB open・verb 実行・表示・nudge 配送)に分解する。
全出力(テーブル・JSON・エラー文言)はバイト不変(msg golden 全 8 green)。

- peermsg.Request(parse 済み正規形)+ Deps(旧 package var seam 4 本
  の構造体化: DetectIdentity/ListLivePanes/SendLine/LoadState)+ Run
- レポート型は JSON タグ・フィールド順・omitempty を verbatim 移送
- rePlanTaskID の正規表現を team.TaskIDRE に一本化(cmd と peermsg の
  双方から参照、重複定義を解消)
