+++ cmd/fanout/msg_test.go
@@ -1,19 +1,11 @@
 package main
 
 import (
-	"errors"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"strings"
 	"testing"
 
 	"github.com/butaosuinu/fanout/internal/core/exitcode"
 	"github.com/butaosuinu/fanout/internal/infra/log"
-	"github.com/butaosuinu/fanout/internal/infra/msgstore"
-	"github.com/butaosuinu/fanout/internal/infra/state"
-	"github.com/butaosuinu/fanout/internal/infra/team"
-	"github.com/butaosuinu/fanout/internal/infra/tmuxrun"
 )
 
 func msgTestLogger() *log.Logger {
@@ -166,435 +158,5 @@ func TestParseMsgFlags(t *testing.T) {
 				tc.want(t, f)
 			}
 		})
-	}
-}
-
-func TestResolveMsgIdentity(t *testing.T) {
-	detected := team.Identity{
-		Issue:  70,
-		Parent: "68",
-		Pane: state.Pane{
-			IssueNum: 70, Parent: "68", PaneID: "%1", Slug: "msg-cli-surface-70",
-			Agent: "claude", DisplayName: "msg cli", WorktreePath: "/tmp/wt",
-		},
-	}
-	for _, tc := range []struct {
-		name       string
-		flags      msgFlags
-		detect     func() (team.Identity, error)
-		code       exitcode.Code
-		wantSelf   int
-		wantParent string
-		wantPane   string // expected pane_id, "" when identity is explicit
-	}{
-		{
-			name:  "both explicit skips detection",
-			flags: msgFlags{verb: "send", self: 70, parent: "68"},
-			detect: func() (team.Identity, error) {
-				t.Error("detect called despite explicit --self/--parent")
-				return team.Identity{}, nil
-			},
-			code: exitcode.OK, wantSelf: 70, wantParent: "68",
-		},
-		{
-			name:   "detection fills both",
-			flags:  msgFlags{verb: "inbox"},
-			detect: func() (team.Identity, error) { return detected, nil },
-			code:   exitcode.OK, wantSelf: 70, wantParent: "68", wantPane: "%1",
-		},
-		{
-			name:   "explicit self keeps detected parent",
-			flags:  msgFlags{verb: "send", self: 99},
-			detect: func() (team.Identity, error) { return detected, nil },
-			code:   exitcode.OK, wantSelf: 99, wantParent: "68",
-		},
-		{
-			name:   "explicit parent keeps detected self",
-			flags:  msgFlags{verb: "send", parent: "77"},
-			detect: func() (team.Identity, error) { return detected, nil },
-			code:   exitcode.OK, wantSelf: 70, wantParent: "77", wantPane: "%1",
-		},
-		{
-			name:  "manual pane negative synthetic issue is accepted",
-			flags: msgFlags{verb: "inbox"},
-			detect: func() (team.Identity, error) {
-				return team.Identity{
-					Issue:  -1,
-					Parent: "@manual",
-					Pane:   state.Pane{IssueNum: -1, Parent: "@manual", PaneID: "%9"},
-				}, nil
-			},
-			code: exitcode.OK, wantSelf: -1, wantParent: "@manual", wantPane: "%9",
-		},
-		{
-			name:   "detection failure",
-			flags:  msgFlags{verb: "send"},
-			detect: func() (team.Identity, error) { return team.Identity{}, team.ErrPaneNotFound },
-			code:   exitcode.Invocation,
-		},
-		{
-			name:   "peers needs only parent",
-			flags:  msgFlags{verb: "peers", parent: "68"},
-			detect: func() (team.Identity, error) { return team.Identity{}, errors.New("must not be called") },
-			code:   exitcode.OK, wantSelf: 0, wantParent: "68",
-		},
-		{
-			name:   "nudge needs only parent",
-			flags:  msgFlags{verb: "nudge", to: 71, parent: "68"},
-			detect: func() (team.Identity, error) { return team.Identity{}, errors.New("must not be called") },
-			code:   exitcode.OK, wantSelf: 0, wantParent: "68",
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			orig := detectIdentity
-			detectIdentity = tc.detect
-			defer func() { detectIdentity = orig }()
-
-			self, parent, pane, code := resolveMsgIdentity(&tc.flags, msgTestLogger())
-			if code != tc.code {
-				t.Fatalf("code = %d, want %d", code, tc.code)
-			}
-			if code != exitcode.OK {
-				return
-			}
-			if self != tc.wantSelf || parent != tc.wantParent {
-				t.Errorf("self, parent = %d, %q, want %d, %q", self, parent, tc.wantSelf, tc.wantParent)
-			}
-			if pane.PaneID != tc.wantPane {
-				t.Errorf("pane.PaneID = %q, want %q", pane.PaneID, tc.wantPane)
-			}
-		})
-	}
-}
-
-func TestWriteMsgMessagesTable(t *testing.T) {
-	to := 70
-	read := "2026-06-13T00:00:00Z"
-	msgs := []msgstore.Message{
-		{ID: 1, From: 71, To: &to, Kind: "note", Body: "multi\nline", CreatedAt: "2026-06-13T00:00:00Z", ReadAt: &read},
-		{ID: 3, From: 71, Board: true, Kind: "note", Body: "board post", CreatedAt: "2026-06-13T00:00:00Z"},
-	}
-	var out strings.Builder
-	lg := log.NewWith(&out, &strings.Builder{}, false)
-	writeMsgMessagesTable(msgs, true, nil, lg)
-	got := out.String()
-	for _, want := range []string{
-		"ID  FROM  TO     KIND  CREATED               BODY",
-		"1   #71   #70    note  2026-06-13T00:00:00Z  multi line",
-		"3   #71   board  note  2026-06-13T00:00:00Z  board post",
-	} {
-		if !strings.Contains(got, want) {
-			t.Errorf("table missing %q in:\n%s", want, got)
-		}
-	}
-
-	out.Reset()
-	writeMsgMessagesTable(nil, false, nil, lg)
-	if got := out.String(); got != "no unread messages\n" {
-		t.Errorf("empty unread table = %q", got)
-	}
-	out.Reset()
-	writeMsgMessagesTable(nil, true, nil, lg)
-	if got := out.String(); got != "no messages\n" {
-		t.Errorf("empty all table = %q", got)
-	}
-}
-
-// TestWriteMsgMessagesTablePlanLabels checks that a plan label map renders
-// synthetic peer numbers as task ids in the FROM/TO columns.
-func TestWriteMsgMessagesTablePlanLabels(t *testing.T) {
-	apiNum := -111
-	dbNum := -222
-	msgs := []msgstore.Message{
-		{ID: 1, From: dbNum, To: &apiNum, Kind: "note", Body: "ping", CreatedAt: "2026-06-13T00:00:00Z"},
-	}
-	labels := map[int]string{apiNum: "api-client", dbNum: "db-layer"}
-	var out strings.Builder
-	lg := log.NewWith(&out, &strings.Builder{}, false)
-	writeMsgMessagesTable(msgs, true, labels, lg)
-	got := out.String()
-	if !strings.Contains(got, "db-layer") || !strings.Contains(got, "api-client") {
-		t.Errorf("plan label table missing task ids in:\n%s", got)
-	}
-	if strings.Contains(got, "#-111") || strings.Contains(got, "#-222") {
-		t.Errorf("plan label table leaked synthetic numbers in:\n%s", got)
-	}
-}
-
-func TestShouldNudge(t *testing.T) {
-	for _, tc := range []struct {
-		state string
-		want  bool
-	}{
-		{"running", true},
-		{"done", false},
-		{"", false},
-		{"idle", false},
-		{"garbage", false},
-	} {
-		if got := shouldNudge(tc.state); got != tc.want {
-			t.Errorf("shouldNudge(%q) = %v, want %v", tc.state, got, tc.want)
-		}
-	}
-}
-
-func TestRunMsgNudge(t *testing.T) {
-	// withPane is a legacy row (no recorded worktree) so it falls back to an
-	// id-only liveness match; withWorktree carries a worktree, so the live pane
-	// must also sit at/under it (the reused-%N defense).
-	withPane := state.Store{SchemaVersion: 1, Panes: []state.Pane{{Parent: "68", IssueNum: 71, PaneID: "%5"}}}
-	withWorktree := state.Store{SchemaVersion: 1, Panes: []state.Pane{{Parent: "68", IssueNum: 71, PaneID: "%5", WorktreePath: "/wt/recipient"}}}
-	noPaneID := state.Store{SchemaVersion: 1, Panes: []state.Pane{{Parent: "68", IssueNum: 72, PaneID: ""}}}
-	lp := func(id, path, agentState string) tmuxrun.LivePane {
-		return tmuxrun.LivePane{ID: id, CurrentPath: path, AgentState: agentState}
-	}
-
-	for _, tc := range []struct {
-		name           string
-		flags          msgFlags
-		store          state.Store
-		storeErr       error
-		live           []tmuxrun.LivePane
-		listErr        error
-		sendErr        error
-		wantCode       exitcode.Code
-		wantListed     bool // listLivePanes consulted
-		wantSendCalled bool // sendLiteralLine invoked
-		wantStdout     string
-		wantStderr     string
-	}{
-		{
-			name: "running pane is nudged (legacy id-only match)", flags: msgFlags{verb: "nudge", to: 71}, store: withPane,
-			live: []tmuxrun.LivePane{lp("%5", "/anywhere", "running")}, wantCode: exitcode.OK, wantListed: true, wantSendCalled: true, wantStdout: "nudged #71",
-		},
-		{
-			name: "running pane at the recorded worktree is nudged", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient", "running")}, wantCode: exitcode.OK, wantListed: true, wantSendCalled: true, wantStdout: "nudged #71",
-		},
-		{
-			name: "running pane under the recorded worktree is nudged", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient/nested", "running")}, wantCode: exitcode.OK, wantListed: true, wantSendCalled: true, wantStdout: "nudged #71",
-		},
-		{
-			// The core Codex P2: tmux reused %5 for a pane sitting elsewhere.
-			// It must NOT be nudged even though it reports "running".
-			name: "reused id off the recorded worktree is not nudged", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/someone-else", "running")}, wantCode: exitcode.OK, wantListed: true, wantStderr: "gone or its id was reused",
-		},
-		{
-			name: "pane absent from the live set is a no-op success", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%9", "/wt/recipient", "running")}, wantCode: exitcode.OK, wantListed: true, wantStderr: "gone or its id was reused",
-		},
-		{
-			// parent "0068" must still resolve the stored "68" pane via Find's
-			// numeric canonicalization (parentMatches).
-			name: "leading-zero parent still resolves the recipient", flags: msgFlags{verb: "nudge", to: 71, parent: "0068"}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient", "running")}, wantCode: exitcode.OK, wantListed: true, wantSendCalled: true, wantStdout: "nudged #71",
-		},
-		{
-			name: "done pane is a no-op success", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient", "done")}, wantCode: exitcode.OK, wantListed: true, wantStderr: "agent is not running",
-		},
-		{
-			name: "unset state is a no-op success", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient", "")}, wantCode: exitcode.OK, wantListed: true, wantStderr: "agent is not running",
-		},
-		{
-			name: "tmux unavailable is a no-op success", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			listErr: errors.New("tmux down"), wantCode: exitcode.OK, wantListed: true, wantStderr: "tmux is unavailable",
-		},
-		{
-			name: "send-keys failure stays a best-effort success", flags: msgFlags{verb: "nudge", to: 71}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient", "running")}, sendErr: errors.New("boom"), wantCode: exitcode.OK, wantListed: true, wantSendCalled: true, wantStderr: "send-keys failed",
-		},
-		{
-			name: "recipient absent from state is a no-op success", flags: msgFlags{verb: "nudge", to: 99}, store: withWorktree,
-			wantCode: exitcode.OK, wantStderr: "not recorded",
-		},
-		{
-			name: "recipient without a recorded pane is a no-op success", flags: msgFlags{verb: "nudge", to: 72}, store: noPaneID,
-			wantCode: exitcode.OK, wantStderr: "no recorded pane",
-		},
-		{
-			name: "state load failure is an invocation error", flags: msgFlags{verb: "nudge", to: 71}, storeErr: errors.New("bad path"),
-			wantCode: exitcode.Invocation,
-		},
-		{
-			name: "json reports a delivered nudge", flags: msgFlags{verb: "nudge", to: 71, json: true}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient", "running")}, wantCode: exitcode.OK, wantListed: true, wantSendCalled: true,
-			wantStdout: `"nudged": true`,
-		},
-		{
-			name: "json reports a skipped nudge with a reason", flags: msgFlags{verb: "nudge", to: 71, json: true}, store: withWorktree,
-			live: []tmuxrun.LivePane{lp("%5", "/wt/recipient", "done")}, wantCode: exitcode.OK, wantListed: true, wantStdout: `"nudged": false`,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			origLoad, origList, origSend := loadStateStore, listLivePanes, sendLiteralLine
-			defer func() { loadStateStore, listLivePanes, sendLiteralLine = origLoad, origList, origSend }()
-
-			loadStateStore = func() (state.Store, error) { return tc.store, tc.storeErr }
-			listed := false
-			listLivePanes = func() ([]tmuxrun.LivePane, error) {
-				listed = true
-				return tc.live, tc.listErr
-			}
-			sent := false
-			var sentPane, sentText string
-			sendLiteralLine = func(paneID, text string) error {
-				sent = true
-				sentPane, sentText = paneID, text
-				return tc.sendErr
-			}
-
-			var out, errb strings.Builder
-			lg := log.NewWith(&out, &errb, false)
-			// resolveMsgIdentity supplies parent in the real flow; mirror that
-			// here, defaulting to "68" so existing cases need no parent field.
-			parent := tc.flags.parent
-			if parent == "" {
-				parent = "68"
-			}
-			code := runMsgNudge(&tc.flags, parent, lg)
-			if code != tc.wantCode {
-				t.Fatalf("runMsgNudge() code = %d, want %d (stderr: %q)", code, tc.wantCode, errb.String())
-			}
-			if tc.wantCode != exitcode.OK {
-				return
-			}
-			if listed != tc.wantListed {
-				t.Errorf("listLivePanes consulted = %v, want %v", listed, tc.wantListed)
-			}
-			if sent != tc.wantSendCalled {
-				t.Errorf("sendLiteralLine called = %v, want %v", sent, tc.wantSendCalled)
-			}
-			if tc.wantSendCalled {
-				if sentPane != "%5" {
-					t.Errorf("send pane = %q, want %%5", sentPane)
-				}
-				if sentText != nudgeText {
-					t.Errorf("send text = %q, want nudgeText", sentText)
-				}
-			}
-			if tc.wantStdout != "" && !strings.Contains(out.String(), tc.wantStdout) {
-				t.Errorf("stdout = %q, want substring %q", out.String(), tc.wantStdout)
-			}
-			if tc.wantStderr != "" && !strings.Contains(errb.String(), tc.wantStderr) {
-				t.Errorf("stderr = %q, want substring %q", errb.String(), tc.wantStderr)
-			}
-		})
-	}
-}
-
-// TestLoadStateStoreResolvesOwnerFromChildWorktree guards the regression Codex
-// flagged: nudge is run from a child worktree pane, whose own git toplevel has
-// no state.json. loadStateStore must climb to the owner (OwnerProjectRoot) that
-// holds the row — resolveStateRuntime would load the child's empty store and
-// report every recipient "not recorded".
-func TestLoadStateStoreResolvesOwnerFromChildWorktree(t *testing.T) {
-	if _, err := exec.LookPath("git"); err != nil {
-		t.Skip("git not installed")
-	}
-	owner := t.TempDir()
-	gitCmdTest(t, owner, "init", "-b", "main")
-	gitCmdTest(t, owner, "config", "user.email", "fanout@example.invalid")
-	gitCmdTest(t, owner, "config", "user.name", "fanout test")
-	if err := os.WriteFile(filepath.Join(owner, "README.md"), []byte("base\n"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	gitCmdTest(t, owner, "add", "README.md")
-	gitCmdTest(t, owner, "commit", "-m", "base")
-
-	child := filepath.Join(owner, ".fanout", "worktrees", "s-71")
-	gitCmdTest(t, owner, "worktree", "add", "-b", "s-71", child)
-
-	// The recipient row lives in the OWNER's state.json, never the child's.
-	if err := os.MkdirAll(filepath.Dir(state.Path(owner)), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	fixture := `{"schemaVersion":1,"panes":[{"parent":"68","issueNum":71,"slug":"s-71","paneId":"%5"}]}` + "\n"
-	if err := os.WriteFile(state.Path(owner), []byte(fixture), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	// Run from the child pane with no FANOUT_STATE_PATH override (the real
-	// in-pane workflow). resolveStateRuntime would resolve the child toplevel
-	// and find nothing.
-	t.Chdir(child)
-	t.Setenv(fanoutStatePathEnv, "")
-
-	st, err := loadStateStore()
-	if err != nil {
-		t.Fatalf("loadStateStore() failed: %v", err)
-	}
-	if _, ok := st.Find("68", 71); !ok {
-		t.Fatalf("loadStateStore() found %d panes, want the owner's recipient #71 (child-worktree owner resolution regressed)", len(st.Panes))
-	}
-}
-
-func TestMatchLivePane(t *testing.T) {
-	panes := []tmuxrun.LivePane{
-		{ID: "%5", CurrentPath: "/wt/recipient", AgentState: "running"},
-		{ID: "%6", CurrentPath: "/wt/recipient/nested/deep", AgentState: "done"},
-		{ID: "%7", CurrentPath: "/wt/other", AgentState: "running"},
-	}
-	for _, tc := range []struct {
-		name     string
-		paneID   string
-		worktree string
-		wantOK   bool
-		wantID   string // matched pane id when wantOK
-	}{
-		{name: "id + exact worktree", paneID: "%5", worktree: "/wt/recipient", wantOK: true, wantID: "%5"},
-		{name: "id + path under worktree", paneID: "%6", worktree: "/wt/recipient", wantOK: true, wantID: "%6"},
-		{name: "trailing slash on worktree still matches", paneID: "%5", worktree: "/wt/recipient/", wantOK: true, wantID: "%5"},
-		{name: "reused id off the worktree is rejected", paneID: "%7", worktree: "/wt/recipient", wantOK: false},
-		{name: "sibling-prefix path is not under the worktree", paneID: "%5", worktree: "/wt/recip", wantOK: false},
-		{name: "id absent from the live set", paneID: "%9", worktree: "/wt/recipient", wantOK: false},
-		{name: "empty worktree falls back to id-only", paneID: "%7", worktree: "", wantOK: true, wantID: "%7"},
-		{name: "empty pane id never matches", paneID: "", worktree: "/wt/recipient", wantOK: false},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			got, ok := matchLivePane(panes, tc.paneID, tc.worktree)
-			if ok != tc.wantOK {
-				t.Fatalf("matchLivePane(%q, %q) ok = %v, want %v", tc.paneID, tc.worktree, ok, tc.wantOK)
-			}
-			if ok && got.ID != tc.wantID {
-				t.Errorf("matched pane id = %q, want %q", got.ID, tc.wantID)
-			}
-		})
-	}
-}
-
-// TestResolveMemberNum covers the seam that decides whether a token is an issue
-// number or a plan task id once the parent is known — in particular that an
-// all-digit token routes to a task under a plan parent (not the issue number).
-func TestResolveMemberNum(t *testing.T) {
-	var buf strings.Builder
-	lg := log.NewWith(&buf, &buf, false)
-	const planParent = "plan:demo"
-
-	n, code := resolveMemberNum("send", "--to", "123", 123, planParent, lg)
-	if code != exitcode.OK {
-		t.Fatalf("plan all-digit code = %d, want OK", code)
-	}
-	if n == 123 || n >= 0 {
-		t.Errorf("plan all-digit resolved to %d, want a negative task number (not the issue number 123)", n)
-	}
-	if again, _ := resolveMemberNum("send", "--to", "123", 123, planParent, lg); again != n {
-		t.Errorf("resolveMemberNum not deterministic: %d != %d", again, n)
-	}
-
-	if _, code := resolveMemberNum("send", "--to", "api-client", 0, "68", lg); code != exitcode.Invocation {
-		t.Error("a task id under an issue parent must be rejected")
-	}
-	if got, code := resolveMemberNum("send", "--to", "71", 71, "68", lg); code != exitcode.OK || got != 71 {
-		t.Errorf("numeric under issue = (%d, %d), want (71, OK)", got, code)
-	}
-	if got, code := resolveMemberNum("send", "--to", "-1", -1, "68", lg); code != exitcode.OK || got != -1 {
-		t.Errorf("manual negative under issue = (%d, %d), want (-1, OK)", got, code)
-	}
-	if _, code := resolveMemberNum("send", "--to", "-1", -1, planParent, lg); code != exitcode.Invocation {
-		t.Error("a non-task token under a plan parent must be rejected")
 	}
 }
